Fix ADPluginBar commonplugins.png name collision
When ADPluginBar was added to the areaDetector repo in commit 7ada691,
it introduced a case-insensitive file system name collision for the file
commonplugins.png because commonPlugins.png (uppercase 'P') already
existed.  This worked on a case-sensitive file system, but it doesn't
work on a case-insensitive file system such as APFS on macOS which
produces the following Git warning message when cloning the repo:

  warning: the following paths have collided (e.g. case-sensitive paths
  on a case-insensitive filesystem) and only one from the same
  colliding group is in the working tree:

    'master/_images/commonPlugins.png'
    'master/_images/commonplugins.png'

To fix this, rename commonplugins.png to ADPluginBar_commonPlugins.png.
The file name has been prefixed with the name of the plugin that the
file belongs to, and the first 'p' has been made uppercase to match the
uppercase 'P' in the existing file commonPlugins.png.